refactor: split `validation.mjs`
