+++ src/commands/cat.mjs
@@ -5,7 +5,7 @@ import { pipeline } from 'node:stream/promises';
 
 import { pathToFile } from '../parameters/index.mjs';
 import { getCommandUsage } from '../utils/commandUsage.mjs';
-import { validatePath } from '../utils/validation.mjs';
+import { validatePath } from '../validation/index.mjs';
 
 const parameters = {
   mandatory: [pathToFile],
