+++ src/commands/decompress.mjs
@@ -5,7 +5,7 @@ import { createBrotliDecompress } from 'node:zlib';
 
 import { pathToFile, pathToDestination } from '../parameters/index.mjs';
 import { getCommandUsage } from '../utils/commandUsage.mjs';
-import { validatePath } from '../utils/validation.mjs';
+import { validatePath } from '../validation/index.mjs';
 
 const parameters = {
   mandatory: [pathToFile, pathToDestination],
