+++ src/commands/rn.mjs
@@ -3,7 +3,7 @@ import { join } from 'node:path';
 
 import { pathToFile, newFileName } from '../parameters/index.mjs';
 import { getCommandUsage } from '../utils/commandUsage.mjs';
-import { validatePath } from '../utils/validation.mjs';
+import { validatePath } from '../validation/index.mjs';
 
 const parameters = {
   mandatory: [pathToFile, newFileName],
